fix: entity, service methods & some tests

+++ src/main/java/com/urfu/library/model/BookRepository.java
@@ -5,7 +5,13 @@ import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
+/**
+ * Репозиторий сущности Book для взаимодействия с базой данных
+ */
 @Repository
-public interface BookRepository extends JpaRepository<Book, Integer> {
+public interface BookRepository extends JpaRepository<Book, Long> {
+    /**
+     * Поиск книг по заданному автору
+     */
     List<Book> findByTitle(String author);
 }
